Use kill to close network streams
Couldn't get vlc to gracefully close when the port closed on the rpi2.
Worked fine on the host, but not on the target rpi2.

This will leave zombie vlc processes on the host if the BEAM exists
abruptly.  On targets, this relies on the Nerves system rebooting if the
BEAM exists.